added description and methods

+++ src/utils/calendarEntryAPI.js
@@ -2,10 +2,16 @@ const BACKEND_URL = "https://big-brother-be-3d6ad173758c.herokuapp.com/";
 
 // Create an item in DynamoDB
 export const createCalendarEntryInDynamoDB = async (item) => {
+    const token = localStorage.getItem('token');
+
+    if (!token) {
+        throw new Error("No token found");
+    }
     try {
         const response = await fetch(`${BACKEND_URL}v1/calendar-entry`, {
             method: "POST",
             headers: {
+                'Authorization': `Bearer ${token}`, 
                 "Content-Type": "application/json",
             },
             body: JSON.stringify(item),
@@ -25,6 +31,11 @@ export const createCalendarEntryInDynamoDB = async (item) => {
 
 // Retrieve an item from DynamoDB
 export const retrieveCalendarEntryFromDynamoDB = async (item) => {
+    const token = localStorage.getItem('token');
+
+    if (!token) {
+        throw new Error("No token found");
+    }
     try {
         const response = await fetch(`${BACKEND_URL}v1/calendar-entry/${item.id}`, {
             method: "GET",
@@ -44,10 +55,16 @@ export const retrieveCalendarEntryFromDynamoDB = async (item) => {
 
 // Update an item in DynamoDB
 export const updateCalendarEntryInDynamoDB = async (item) => {
+    const token = localStorage.getItem('token');
+
+    if (!token) {
+        throw new Error("No token found");
+    }
     try {
         const response = await fetch(`${BACKEND_URL}v1/calendar-entry/${item.id}`, {
             method: "PUT",
             headers: {
+                'Authorization': `Bearer ${token}`, 
                 "Content-Type": "application/json",
             },
             body: JSON.stringify(item),
@@ -67,10 +84,16 @@ export const updateCalendarEntryInDynamoDB = async (item) => {
 
 // Delete an item from DynamoDB
 export const deleteCalendarEntryFromDynamoDB = async (item) => {
-  try {
+    const token = localStorage.getItem('token');
+
+    if (!token) {
+        throw new Error("No token found");
+    }
+    try {
     const response = await fetch(`${BACKEND_URL}v1/calendar-entry/${item.id}`, {
       method: "DELETE",
       headers: {
+        'Authorization': `Bearer ${token}`, 
         "Content-Type": "application/json",
       },
     });
